requirements.txt
mss
Pillow
numpy
pynput
